reviewing whole project, moving alll types into the @pnpm/types package

+++ pkg-manager/modules-cleaner/src/removeDirectDependency.ts
@@ -1,9 +1,11 @@
-import path from 'path'
-import { promises as fs } from 'fs'
-import { rootLogger } from '@pnpm/core-loggers'
-import { removeBin, removeBinsOfDependency } from '@pnpm/remove-bins'
-import { type DependenciesField } from '@pnpm/types'
+import path from 'node:path'
+import { promises as fs } from 'node:fs'
+
 import rimraf from '@zkochan/rimraf'
+
+import { rootLogger } from '@pnpm/core-loggers'
+import type { DependenciesField } from '@pnpm/types'
+import { removeBin, removeBinsOfDependency } from '@pnpm/remove-bins'
 
 export async function removeDirectDependency(
   dependency: {
@@ -11,18 +13,20 @@ export async function removeDirectDependency(
     name: string
   },
   opts: {
-    binsDir: string
-    dryRun?: boolean
+    binsDir?: string | undefined
+    dryRun?: boolean | undefined
     modulesDir: string
-    muteLogs?: boolean
+    muteLogs?: boolean | undefined
     rootDir: string
   }
 ) {
   const dependencyDir = path.join(opts.modulesDir, dependency.name)
+
   const results = await Promise.all([
     removeBinsOfDependency(dependencyDir, opts),
     !opts.dryRun && (removeBin(dependencyDir) as any), // eslint-disable-line @typescript-eslint/no-explicit-any
   ])
+
   await removeIfEmpty(opts.binsDir)
 
   const uninstalledPkg = results[0]
@@ -43,15 +47,16 @@ export async function removeDirectDependency(
   }
 }
 
-export async function removeIfEmpty(dir: string) {
-  if (await dirIsEmpty(dir)) {
+export async function removeIfEmpty(dir: string | undefined): Promise<void> {
+  if (dir && await dirIsEmpty(dir)) {
     await rimraf(dir)
   }
 }
 
-async function dirIsEmpty(dir: string) {
+async function dirIsEmpty(dir: string): Promise<boolean> {
   try {
     const fileNames = await fs.readdir(dir)
+
     return fileNames.length === 0
   } catch {
     return false
